backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pandas
openpyxl
python-multipart
python-dotenv
